Add pytest fixtures for testing AIApp and refactor test_config.py
